Password String Added to User
Honestly, just wanted to get something out there before the three days passed. If this is actually superfluous then I'll leave it to languish here. As is, figured I would get this added in. Added the getter/setter and modified all 'test cases' in which User() was called since I had to edit that.